Move EnrichedEvent to endpoint_standard, fix underlying Query class name and check for where criteria in unit test
