+++ src/cbc_sdk/endpoint_standard/__init__.py
@@ -2,4 +2,4 @@
 
 from __future__ import absolute_import
 
-from cbc_sdk.endpoint_standard.base import Device, Event, Policy, Query
+from cbc_sdk.endpoint_standard.base import Device, Event, Policy, EnrichedEvent, Query, AsyncEnrichedEventQuery
