+++ src/cbc_sdk/enterprise_edr/__init__.py
@@ -1,7 +1,7 @@
 from __future__ import absolute_import
 
-from cbc_sdk.enterprise_edr.base import (Process, Event, Tree, Query, AsyncProcessQuery, 
-                                        EnrichedEvent, AsyncEventQuery, TreeQuery)
+from cbc_sdk.enterprise_edr.base import (Process, Event, Tree, Query, 
+                                        AsyncProcessQuery, TreeQuery)
 
 from cbc_sdk.enterprise_edr.threat_intelligence import (Watchlist, Feed, Report,
                                                       ReportSeverity, IOC, IOC_V2,
